refactor: rename telegram_update
