+++ src/main/java/com/bot/spider/dtos/telegram/CallbackQueryDTO.java
@@ -1,4 +1,4 @@
-package com.bot.spider.dtos.telegram_update;
+package com.bot.spider.dtos.telegram;
 
 public record CallbackQueryDTO(
         String id,
