+++ src/main/java/com/bot/spider/dtos/telegram/FileDTO.java
@@ -1,4 +1,4 @@
-package com.bot.spider.dtos.telegram_update;
+package com.bot.spider.dtos.telegram;
 
 public record FileDTO(
         String file_id,
