+++ src/main/java/com/bot/spider/dtos/telegram/MessageDTO.java
@@ -1,4 +1,4 @@
-package com.bot.spider.dtos.telegram_update;
+package com.bot.spider.dtos.telegram;
 
 import java.util.Optional;
