+++ src/main/java/com/bot/spider/services/InputProcess/ExampleHandlingInputs.java
@@ -1,6 +1,6 @@
 package com.bot.spider.services.InputProcess;
 
-import com.bot.spider.dtos.telegram_update.TelegramUpdateDTO;
+import com.bot.spider.dtos.telegram.TelegramUpdateDTO;
 import com.bot.spider.enums.TelegramChatAction;
 import com.bot.spider.libs.keyboard.CreateKeyboard;
 import com.bot.spider.libs.keyboard.InlineKeyboard;
